+++ valgrind-requests/CHANGELOG.md
@@ -18,6 +18,9 @@ and this project adheres to
 
 ## [Unreleased]
 
+- Extract valgrind support table from build script into new gungraun-common
+  crate.
+
 ## [1.1.0] - 2026-05-14
 
 ### Added